+++ example/lib/main.dart
@@ -5,13 +5,13 @@ import 'next_page.dart';
 
 void main() {
   FeaturesTour.setGlobalConfig(
-    force: true,
+    // force: true,
     skipConfig: SkipConfig.copyWith(
       text: 'SKIP >>>',
       buttonStyle: TextButton.styleFrom(
         shape: RoundedRectangleBorder(
           side: const BorderSide(
-            color: Colors.red,
+            color: Colors.blue,
           ),
           borderRadius: BorderRadius.circular(10),
         ),
@@ -19,10 +19,19 @@ void main() {
     ),
     nextConfig: NextConfig.copyWith(
       text: 'NEXT >>',
+      buttonStyle: TextButton.styleFrom(
+        shape: RoundedRectangleBorder(
+          side: const BorderSide(
+            color: Colors.blue,
+          ),
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
     ),
     predialogConfig: PredialogConfig.copyWith(
       enabled: true,
     ),
+    debugLog: true,
   );
 
   runApp(const MaterialApp(home: MyApp()));